1.8beta

+++ cn/lambdacraft/deathmatch/entity/fx/EntityGaussRay.java
@@ -57,22 +57,22 @@ public class EntityGaussRay extends Entity {
 	public void setRayHeading(double par1, double par3, double par5,
 			float par7, float par8) {
         float f2 = MathHelper.sqrt_double(par1 * par1 + par3 * par3 + par5 * par5);
-        par1 /= (double)f2;
-        par3 /= (double)f2;
-        par5 /= (double)f2;
-        par1 += this.rand.nextGaussian() * 0.007499999832361937D * (double)par8;
-        par3 += this.rand.nextGaussian() * 0.007499999832361937D * (double)par8;
-        par5 += this.rand.nextGaussian() * 0.007499999832361937D * (double)par8;
-        par1 *= (double)par7;
-        par3 *= (double)par7;
-        par5 *= (double)par7;
+        par1 /= f2;
+        par3 /= f2;
+        par5 /= f2;
+        par1 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+        par3 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+        par5 += this.rand.nextGaussian() * 0.007499999832361937D * par8;
+        par1 *= par7;
+        par3 *= par7;
+        par5 *= par7;
         this.motionX = par1;
         this.motionY = par3;
         this.motionZ = par5;
         float f3 = MathHelper.sqrt_double(par1 * par1 + par5 * par5);
         //System.out.println("(" + par1 + ", " + par3 + ", " + par5 + ")");
         this.prevRotationYaw = this.rotationYaw = (float)(Math.atan2(par1, par5) * 180.0D / Math.PI);
-        this.prevRotationPitch = this.rotationPitch = (float)(Math.atan2(par3, (double)f3) * 180.0D / Math.PI);
+        this.prevRotationPitch = this.rotationPitch = (float)(Math.atan2(par3, f3) * 180.0D / Math.PI);
        // System.out.println(rotationYaw + " " + rotationPitch);
 	}
