+++ cn/lambdacraft/terrain/world/chunk/WorldChunkManagerXenContinent.java
@@ -89,7 +89,7 @@ public class WorldChunkManagerXenContinent extends WorldChunkManager {
 
         for (int i1 = 0; i1 < par4 * par5; ++i1)
         {
-            float f = (float)BiomeGenBase.getBiome(aint[i1]).getIntRainfall() / 65536.0F;
+            float f = BiomeGenBase.getBiome(aint[i1]).getIntRainfall() / 65536.0F;
 
             if (f > 1.0F)
             {
